Move ReplayGain class from util/ to track/
