+++ src/track/replaygain.cpp
@@ -1,4 +1,4 @@
-#include "util/replaygain.h"
+#include "track/replaygain.h"
 
 #include "util/math.h"
